Refactor code for tests
