+++ app/cart/CheckoutButton.js
@@ -1,0 +1,31 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+
+export default function CheckoutButton() {
+  const router = useRouter();
+
+  return (
+    <button
+      className="checkout-button"
+      data-test-id="cart-checkout"
+      onClick={async (event) => {
+        event.preventDefault();
+        await router.push('/checkout');
+      }}
+    >
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        height="1.6rem"
+        viewBox="0 -960 960 960"
+        width="1.6rem"
+        fill="#FFFFFF"
+      >
+        <path d="m480-560-56-56 63-64H320v-80h167l-64-64 57-56 160 160-160 160ZM280-80q-33 0-56.5-23.5T200-160q0-33 23.5-56.5T280-240q33 0 56.5 23.5T360-160q0 33-23.5 56.5T280-80Zm400 0q-33 0-56.5-23.5T600-160q0-33 23.5-56.5T680-240q33 0 56.5 23.5T760-160q0 33-23.5 56.5T680-80ZM40-800v-80h131l170 360h280l156-280h91L692-482q-11 20-29.5 31T622-440H324l-44 80h480v80H280q-45 0-68.5-39t-1.5-79l54-98-144-304H40Z" />
+      </svg>
+      <span>
+        <strong>Checkout</strong>
+      </span>
+    </button>
+  );
+}
