+++ app/cart/page.js
@@ -1,12 +1,12 @@
 import './page.scss';
 import Image from 'next/image';
 import Link from 'next/link';
-import { redirect } from 'next/navigation';
 import { getSingleProductInsecure } from '../../database/products';
 import { getCookie } from '../../util/cookies';
 import { parseJson } from '../../util/json';
+import { totalCountCalc } from '../../util/product-calculator';
 import CartQuantity from '../common/CartQuantity/CartQuantity';
-import { findProductInCookieCart } from './actions';
+import CheckoutButton from './CheckoutButton';
 
 export function generateMetadata() {
   return {
@@ -16,9 +16,9 @@ export function generateMetadata() {
 }
 
 export default async function Cart() {
-  const cookieCart = getCookie('cart');
+  const cookieCart = await getCookie('cart');
   const cart = !cookieCart ? [] : parseJson(cookieCart);
-  const productsInCart = await findProductInCookieCart();
+  const productsInCart = await totalCountCalc();
 
   const cartTotalValue = productsInCart.reduce((acc, product) => {
     return acc + Number(product.price) * Number(product.count);
@@ -161,26 +161,9 @@ export default async function Cart() {
               </p>{' '}
               €
             </div>
-            <button
-              className="checkout-link"
-              data-test-id="cart-checkout"
-              onClick={async () => {
-                await redirect('/checkout');
-              }}
-            >
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                height="1.6rem"
-                viewBox="0 -960 960 960"
-                width="1.6rem"
-                fill="#FFFFFF"
-              >
-                <path d="m480-560-56-56 63-64H320v-80h167l-64-64 57-56 160 160-160 160ZM280-80q-33 0-56.5-23.5T200-160q0-33 23.5-56.5T280-240q33 0 56.5 23.5T360-160q0 33-23.5 56.5T280-80Zm400 0q-33 0-56.5-23.5T600-160q0-33 23.5-56.5T680-240q33 0 56.5 23.5T760-160q0 33-23.5 56.5T680-80ZM40-800v-80h131l170 360h280l156-280h91L692-482q-11 20-29.5 31T622-440H324l-44 80h480v80H280q-45 0-68.5-39t-1.5-79l54-98-144-304H40Z" />
-              </svg>
-              <span>
-                <strong>Checkout</strong>
-              </span>
-            </button>
+            <form>
+              <CheckoutButton />{' '}
+            </form>
           </div>
         </div>
       )}
