+++ app/common/CartQuantity/actions.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import { createOrUpdateCookie } from '../AddToCart/actions';
+import { createOrUpdateCookie } from '../../../util/cookies';
 
 export default async function UpdateCartCount(id: number, cartCount: number) {
   return await createOrUpdateCookie(id, cartCount);
