--- util/__tests__/createCookies.test.ts
@@ -1,12 +1,0 @@
-// import { expect, test } from '@jest/globals';
-// import { createOrUpdateCookie } from '../../app/common/AddToCart/actions';
-// import { getCookie } from '../cookies';
-
-// test('should set cookie', async () => {
-//   // execute actual logic
-//   await createOrUpdateCookie(1, 1);
-//   // retrieve the result
-//   const resultCookie = cgetCookie('cart');
-//   // expects here
-//   expect(resultCookie['cart']).toEqual('[{"productId":1,"count":1}]');
-// });
